Changes:
- move monkay to _monkay.py file
- deprecated EnvironmentType
- remove edgy.conf.functional. Its only use was the settings.
